@@ -10,98 +10,101 @@
   ],
   "license": "Apache-2.0",
   "scripts": {
-    "start": "webpack-dev-server --progress --colors --config webpack.dev.js",
+    "start": "webpack serve --progress --config webpack.dev.js",
     "prebuild": "yarn run test",
-    "build": "webpack -p --config webpack.prod.js",
+    "build": "webpack --mode production --config webpack.prod.js",
     "test": "jest --reporters=jest-standard-reporter",
     "prettier": "prettier --write \"**/*.{{c,le,sc}ss,g?(raph)ql,htm?(l),js?(on|on5|onl|x|s),md?(x|wn),m?(ark)down,mkdn,ts?(x),vue,y?(a)ml}\"",
     "lint": "eslint 'src/**/*.{ts,tsx,js}'",
     "fix": "eslint 'src/**/*.{ts,tsx,js}' --fix && import-sort --write 'src/**/*.{ts,tsx,js}' && yarn run prettier"
   },
   "devDependencies": {
-    "@babel/cli": "^7.6.4",
-    "@babel/core": "^7.6.4",
-    "@babel/plugin-proposal-class-properties": "^7.5.5",
-    "@babel/plugin-proposal-decorators": "^7.6.0",
-    "@babel/plugin-proposal-do-expressions": "^7.6.0",
-    "@babel/plugin-proposal-export-default-from": "^7.5.2",
-    "@babel/plugin-proposal-export-namespace-from": "^7.5.2",
-    "@babel/plugin-proposal-function-bind": "^7.2.0",
-    "@babel/plugin-proposal-function-sent": "^7.5.0",
-    "@babel/plugin-proposal-json-strings": "^7.2.0",
-    "@babel/plugin-proposal-logical-assignment-operators": "^7.2.0",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.4.4",
-    "@babel/plugin-proposal-numeric-separator": "^7.2.0",
-    "@babel/plugin-proposal-optional-chaining": "^7.6.0",
-    "@babel/plugin-proposal-pipeline-operator": "^7.5.0",
-    "@babel/plugin-proposal-private-methods": "^7.6.0",
-    "@babel/plugin-proposal-throw-expressions": "^7.2.0",
-    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/plugin-syntax-import-meta": "^7.2.0",
-    "@babel/plugin-transform-runtime": "^7.6.2",
-    "@babel/preset-env": "^7.6.3",
-    "@babel/preset-react": "^7.6.3",
+    "@babel/cli": "^7.13.16",
+    "@babel/core": "^7.18.5",
+    "@babel/eslint-parser": "^7.18.2",
+    "@babel/plugin-proposal-class-properties": "^7.13.0",
+    "@babel/plugin-proposal-decorators": "^7.13.15",
+    "@babel/plugin-proposal-do-expressions": "^7.12.13",
+    "@babel/plugin-proposal-export-default-from": "^7.12.13",
+    "@babel/plugin-proposal-export-namespace-from": "^7.12.13",
+    "@babel/plugin-proposal-function-bind": "^7.12.13",
+    "@babel/plugin-proposal-function-sent": "^7.12.13",
+    "@babel/plugin-proposal-json-strings": "^7.13.8",
+    "@babel/plugin-proposal-logical-assignment-operators": "^7.13.8",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.13.8",
+    "@babel/plugin-proposal-numeric-separator": "^7.12.13",
+    "@babel/plugin-proposal-optional-chaining": "^7.13.12",
+    "@babel/plugin-proposal-pipeline-operator": "^7.12.13",
+    "@babel/plugin-proposal-private-methods": "^7.13.0",
+    "@babel/plugin-proposal-throw-expressions": "^7.12.13",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/plugin-syntax-import-meta": "^7.10.4",
+    "@babel/plugin-transform-runtime": "^7.18.5",
+    "@babel/preset-env": "^7.13.15",
+    "@babel/preset-react": "^7.13.13",
     "babel-eslint": "^10.0.3",
-    "babel-jest": "^24.9.0",
-    "babel-loader": "^8.0.6",
-    "copy-webpack-plugin": "^5.0.4",
-    "css-loader": "^3.2.0",
-    "eslint": "^6.6.0",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-config-prettier": "^6.5.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-jest": "^23.0.2",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-prettier": "^3.1.1",
-    "eslint-plugin-react": "^7.16.0",
-    "eslint-plugin-react-hooks": "^1.7.0",
-    "fibers": "^4.0.2",
-    "file-loader": "^4.2.0",
-    "html-webpack-plugin": "^3.2.0",
+    "babel-jest": "^28.1.1",
+    "babel-loader": "^8.2.2",
+    "copy-webpack-plugin": "^11.0.0",
+    "css-loader": "^6.7.1",
+    "eslint": "^8.17.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^8.2.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jest": "^26.5.3",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.23.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "fibers": "^5.0.0",
+    "file-loader": "^6.2.0",
+    "html-webpack-plugin": "^5.3.1",
     "import-sort-cli": "^6.0.0",
     "import-sort-parser-babylon": "^6.0.0",
     "import-sort-style-renke": "^6.0.0",
-    "jest-cli": "^24.9.0",
-    "jest-fetch-mock": "^2.1.2",
-    "jest-standard-reporter": "^1.0.2",
-    "node-sass": "^4.13.0",
-    "prettier": "^1.18.2",
-    "react-test-renderer": "^16.11.0",
-    "sass": "^1.3.0",
-    "sass-loader": "^8.0.0",
-    "style-loader": "^1.0.0",
-    "typescript": "^3.7.2",
-    "webpack": "^4.41.2",
-    "webpack-bundle-analyzer": "^3.6.0",
+    "jest": "^28.1.1",
+    "jest-cli": "^28.1.1",
+    "jest-fetch-mock": "^3.0.3",
+    "jest-standard-reporter": "^2.0.0",
+    "node-sass": "^7.0.1",
+    "prettier": "^2.7.1",
+    "react-test-renderer": "^18.2.0",
+    "sass": "^1.52.3",
+    "sass-loader": "^13.0.0",
+    "style-loader": "^3.3.1",
+    "typescript": "^4.7.3",
+    "webpack": "^5.73.0",
+    "webpack-bundle-analyzer": "^4.4.1",
     "webpack-bundle-size-analyzer": "^3.1.0",
-    "webpack-cli": "^3.3.10",
-    "webpack-dev-server": "^3.9.0",
-    "webpack-merge": "^4.2.2",
-    "webpack-pwa-manifest": "^4.1.1",
-    "workbox-webpack-plugin": "^4.3.1"
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-server": "^4.9.2",
+    "webpack-merge": "^5.7.3",
+    "webpack-pwa-manifest": "^4.3.0",
+    "workbox-webpack-plugin": "^6.1.5"
   },
   "dependencies": {
-    "@babel/runtime": "^7.6.3",
-    "body-scroll-lock": "^3.0.1",
-    "file-saver": "^2.0.2",
+    "@babel/runtime": "^7.13.17",
+    "body-scroll-lock": "^3.1.5",
+    "file-saver": "^2.0.5",
     "hamburgers": "^1.1.3",
     "immutable": "^4.0.0-rc.12",
-    "js-search": "^1.4.3",
-    "jszip": "^3.2.2",
+    "js-search": "^2.0.0",
+    "jszip": "^3.6.0",
     "lodash.get": "^4.4.2",
     "lodash.set": "^4.3.2",
-    "prism-react-renderer": "^1.0.2",
-    "prismjs": "^1.17.1",
+    "prism-react-renderer": "^1.2.0",
+    "prismjs": "^1.24.0",
     "prop-types": "^15.7.2",
-    "query-string": "^6.8.3",
-    "react": "^16.11.0",
+    "query-string": "^7.0.0",
+    "querystring-es3": "^0.2.1",
+    "react": "^18.2.0",
     "react-body-classname": "^1.3.1",
-    "react-copy-to-clipboard": "^5.0.1",
-    "react-dom": "^16.11.0",
+    "react-copy-to-clipboard": "^5.0.3",
+    "react-dom": "^18.2.0",
     "react-hotkeys": "^2.0.0",
-    "react-markdown": "^4.2.2",
-    "react-toastify": "^5.4.0",
-    "react-transition-group": "^4.3.0"
+    "react-markdown": "^8.0.3",
+    "react-toastify": "^9.0.4",
+    "react-transition-group": "^4.4.1"
   },
   "prettier": {
     "endOfLine": "lf",
