@@ -3,20 +3,22 @@ import React from 'react'
 
 import { IconTimes } from '../icons'
 
-const Close = ({ onClose }) => (
-  <a
-    href='/#'
-    className='toast-close'
-    onClick={event => {
-      event.preventDefault()
-      if (onClose) {
-        onClose()
-      }
-    }}
-  >
-    <IconTimes />
-  </a>
-)
+function Close({ onClose }) {
+  return (
+    <a
+      href='/#'
+      className='toast-close'
+      onClick={event => {
+        event.preventDefault()
+        if (onClose) {
+          onClose()
+        }
+      }}
+    >
+      <IconTimes />
+    </a>
+  )
+}
 
 Close.defaultProps = {
   onClose: null,
