@@ -52,77 +52,73 @@ function Popover({ shareUrl, open, onClose }) {
 
   const urlToShare = `${windowsUtils.origin}/#!${shareUrl}`
   return (
-    <>
-      <TransitionGroup component={null}>
-        {open && (
-          <CSSTransition onEnter={onEnter} classNames='popup' timeout={300}>
-            <div className='popup-share'>
-              <div className='popop-share-container' ref={wrapper}>
-                <div className='popup-header'>
-                  <h1>Share your configuration</h1>
-                </div>
-                <div className='popup-content'>
-                  {/* eslint-disable-next-line */}
+    <TransitionGroup component={null}>
+      {open && (
+        <CSSTransition onEnter={onEnter} classNames='popup' timeout={300}>
+          <div className='popup-share'>
+            <div className='popop-share-container' ref={wrapper}>
+              <div className='popup-header'>
+                <h1>分享你的配置</h1>
+              </div>
+              <div className='popup-content'>
+                {/* eslint-disable-next-line */}
                   <label htmlFor='input-share'>
-                    Use this link to share the current configuration. Attributes
-                    can be removed from the URL if you want to rely on our
-                    defaults.
-                  </label>
-                  <div className='control'>
-                    <input
-                      onFocus={event => {
-                        event.target.select()
-                      }}
-                      id='input-share'
-                      className={`input ${
-                        button === 'Copied!' ? 'padding-lg' : ''
-                      }`}
-                      onKeyDown={event => {
-                        if (event.key === 'Escape') {
-                          onClose()
-                        }
-                      }}
-                      readOnly
-                      value={urlToShare}
-                      ref={input}
-                    />
-                    <CopyToClipboard onCopy={onCopy} text={urlToShare}>
-                      <a
-                        href='/#'
-                        onClick={e => {
-                          e.preventDefault()
-                        }}
-                        className='button'
-                        ref={link}
-                      >
-                        <span className='button-content' tabIndex='-1'>
-                          <span>{button}</span>
-                        </span>
-                      </a>
-                    </CopyToClipboard>
-                  </div>
-                </div>
-                <div className='popup-action'>
-                  <a
-                    href='/#'
-                    onClick={e => {
-                      e.preventDefault()
-                      onClose()
+                  使用此链接可共享当前配置。如果您希望依赖我们的默认值，可以从URL中删除属性。
+                </label>
+                <div className='control'>
+                  <input
+                    onFocus={event => {
+                      event.target.select()
                     }}
-                    className='button'
-                  >
-                    <span className='button-content' tabIndex='-1'>
-                      <span>Close</span>
-                      <span className='secondary desktop-only'>ESC</span>
-                    </span>
-                  </a>
+                    id='input-share'
+                    className={`input ${
+                      button === 'Copied!' ? 'padding-lg' : ''
+                    }`}
+                    onKeyDown={event => {
+                      if (event.key === 'Escape') {
+                        onClose()
+                      }
+                    }}
+                    readOnly
+                    value={urlToShare}
+                    ref={input}
+                  />
+                  <CopyToClipboard onCopy={onCopy} text={urlToShare}>
+                    <a
+                      href='/#'
+                      onClick={e => {
+                        e.preventDefault()
+                      }}
+                      className='button'
+                      ref={link}
+                    >
+                      <span className='button-content' tabIndex='-1'>
+                        <span>{button}</span>
+                      </span>
+                    </a>
+                  </CopyToClipboard>
                 </div>
               </div>
+              <div className='popup-action'>
+                <a
+                  href='/#'
+                  onClick={e => {
+                    e.preventDefault()
+                    onClose()
+                  }}
+                  className='button'
+                >
+                  <span className='button-content' tabIndex='-1'>
+                    <span>Close</span>
+                    <span className='secondary desktop-only'>ESC</span>
+                  </span>
+                </a>
+              </div>
             </div>
-          </CSSTransition>
-        )}
-      </TransitionGroup>
-    </>
+          </div>
+        </CSSTransition>
+      )}
+    </TransitionGroup>
   )
 }
 
