@@ -1,28 +1,31 @@
 /* eslint-disable */
 
-const merge = require('webpack-merge')
+const { merge } = require('webpack-merge')
 const common = require('./webpack.common.js')
 const path = require('path')
 
-const mock = require('./dev/api.json')
+const mock = require('./dev/api.mock.json')
 const fs = require('fs')
 
 const config = {
   mode: 'development',
   devtool: 'inline-source-map',
   devServer: {
-    contentBase: path.resolve(__dirname, 'public'),
+    static: path.resolve(__dirname, 'public'),
     historyApiFallback: true,
     compress: true,
     open: false,
-    before: function(app, server, compiler) {
-      app.get('/metadata/client', function(req, res) {
+    onAfterSetupMiddleware: function (devServer) {
+      if (!devServer) {
+        throw new Error('webpack-dev-server is not defined')
+      }
+      devServer.app.get('/metadata/client', function (req, res) {
         setTimeout(() => {
           res.json(mock)
         }, 800)
       })
-      app.get('/starter.zip', function(req, res) {
-        fs.readFile(path.resolve('./dev/starter.zip'), (err, data) => {
+      devServer.app.get('/starter.zip', function (req, res) {
+        fs.readFile(path.resolve('./dev/starter.mock.zip'), (err, data) => {
           if (err) return sendError(err, res)
           setTimeout(() => {
             res.send(data)
